modified style of file
