+++ commands.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/fatih/color"
 	"github.com/howeyc/gopass"
-	"github.com/upyun/go-sdk/v3/upyun"
 	"github.com/urfave/cli"
 )
 
@@ -480,62 +479,6 @@ func NewUpgradeCommand() cli.Command {
 		Action: func(c *cli.Context) error {
 			Upgrade()
 			return nil
-		},
-	}
-}
-
-//todo
-func NewMvCommand() cli.Command { //文件移动到目录下
-	return cli.Command{
-		Name:  "mv",
-		Usage: "move file",
-		Action: func(c *cli.Context) error {
-			InitAndCheck(LOGIN, CHECK, c)
-			srcpath := c.Args().First() //获得第一个参数文件 src 地址
-			dstpath := c.Args().Get(1)  //获得第二个参数文件 dst 地址
-
-			var isFore int
-			if c.Bool("f") { //强制覆盖
-				isFore = FORCE
-			}
-
-			session.Mv(isFore, &upyun.MoveObjectConfig{
-				SrcPath:  srcpath,
-				DestPath: dstpath,
-			}) //处理移动的逻辑
-
-			return nil
-		},
-		Flags: []cli.Flag{
-			cli.BoolFlag{Name: "f", Usage: "move the file force"},
-		},
-	}
-}
-
-func NewCpCommand() cli.Command {
-	return cli.Command{
-		Name:  "cp",
-		Usage: "copy file",
-		Action: func(c *cli.Context) error {
-			InitAndCheck(LOGIN, CHECK, c)
-			srcpath := c.Args().First()
-			dstpath := c.Args().Get(1)
-
-			var flag int
-			if c.Bool("f") {
-				flag = FORCE
-			}
-
-			session.Cp(flag, &upyun.CopyObjectConfig{
-				SrcPath:  srcpath,
-				DestPath: dstpath,
-			}) //移动逻辑
-
-			return nil
-
-		},
-		Flags: []cli.Flag{
-			cli.BoolFlag{Name: "f", Usage: "copy file force"},
 		},
 	}
 }
