+++ session.go
@@ -977,120 +977,13 @@ func (sess *Session) Purge(urls []string, file string) {
 	}
 }
 
-func (sess *Session) Mv(flag int, conf *upyun.MoveObjectConfig) {
-	//输入的地址为空
-	if len(conf.SrcPath) == 0 || len(conf.DestPath) == 0 {
-		PrintErrorAndExit("File source or destination is not null")
-	}
-
-	var fileName string
-	conf.DestPath, fileName = path.Split(conf.DestPath)
-
-	//判断移动的文件是绝对路径还是相对路径
-	upPath, file := path.Split(conf.SrcPath)
-	conf.SrcPath = path.Join(sess.AbsPath(upPath), file)
-	//判断要移动的文件是否存在
-	_, err := sess.updriver.GetInfo(conf.SrcPath)
-	if err != nil {
-		PrintErrorAndExit("File %s not found", conf.SrcPath)
-	}
-
-	//判断目的目录是否存在
-	conf.DestPath = sess.AbsPath(conf.DestPath)
-	isDir, exist := sess.IsUpYunDir(conf.DestPath)
-	if !isDir {
-		PrintErrorAndExit("%s is not a dir", conf.DestPath)
-	}
-	//如果目的目录不存在，则创建
-	if !exist {
-		if err := sess.updriver.Mkdir(conf.DestPath); err != nil {
-			PrintErrorAndExit("mkdir file error = %s", err)
-		}
-	}
-	//判断输入的目的目录是一个目录还是目录加重命名的文件名
-	if len(fileName) == 0 { //不重命名
-		conf.DestPath = path.Join(conf.DestPath, file)
-	} else {
-		conf.DestPath = path.Join(conf.DestPath, fileName)
-	}
-
-	//判断目的地址是否存在这个文件
-	_, err = sess.updriver.GetInfo(conf.DestPath)
-	//文件存在不覆盖
-	if flag == NOT && err == nil {
-		PrintErrorAndExit("File %s is exist", conf.DestPath)
-	}
-
-	//移动
-	err = sess.updriver.Move(conf)
-	if err != nil {
-		PrintErrorAndExit("MOVE file %s failed", conf.SrcPath)
-	} else {
-		PrintOnlyVerbose("MOVE %s OK", conf.SrcPath)
-		return
-	}
-}
-
-func (sess *Session) Cp(flag int, conf *upyun.CopyObjectConfig) {
-	//输入的地址为空
-	if len(conf.SrcPath) == 0 || len(conf.DestPath) == 0 {
-		PrintErrorAndExit("File source or destination is not null")
-	}
-
-	var fileName string
-	conf.DestPath, fileName = path.Split(conf.DestPath)
-
-	//判断移动的文件是绝对路径还是相对路径
-	upPath, file := path.Split(conf.SrcPath)
-	conf.SrcPath = path.Join(sess.AbsPath(upPath), file)
-	//判断要移动的文件是否存在
-	_, err := sess.updriver.GetInfo(conf.SrcPath)
-	if err != nil {
-		PrintErrorAndExit("File %s not found", conf.SrcPath)
-	}
-
-	//判断目的目录是否存在
-	conf.DestPath = sess.AbsPath(conf.DestPath)
-	isDir, exist := sess.IsUpYunDir(conf.DestPath)
-	if !isDir {
-		PrintErrorAndExit("%s is not a dir", conf.DestPath)
-	}
-	//如果目的目录不存在，则创建
-	if !exist {
-		if err := sess.updriver.Mkdir(conf.DestPath); err != nil {
-			PrintErrorAndExit("mkdir file error = %s", err)
-		}
-	}
-	//判断输入的目的目录是一个目录还是目录加重命名的文件名
-	if len(fileName) == 0 { //不重命名
-		conf.DestPath = path.Join(conf.DestPath, file)
-	} else {
-		conf.DestPath = path.Join(conf.DestPath, fileName)
-	}
-	//判断目的地址是否存在这个文件
-	_, err = sess.updriver.GetInfo(conf.DestPath)
-	//文件存在不覆盖
-	if flag == 0 && err == nil {
-		PrintErrorAndExit("File %s is exist", conf.DestPath)
-	}
-
-	//移动
-	err = sess.updriver.Copy(conf)
-	if err != nil {
-		PrintErrorAndExit("COPY file %s failed", conf.SrcPath)
-	} else {
-		PrintOnlyVerbose("COPY %s OK", conf.SrcPath)
-		return
-	}
-}
-
 // ResumePut upx 命令行实现续传临时信息保存和清理命令 resume-put
 //错误时输出结构体信息
 
-var TempPath = `upyun-resume-recoder.txt`
+var Temp = `upyun-resume-recoder.txt`
 
 func (sess *Session) ResumePut(src, dst string) {
-	_, err := os.OpenFile(src, os.O_WRONLY, 0666)
+	_, err := os.Stat(src)
 	if err != nil {
 		PrintErrorAndExit("file %s may not exist", src)
 	}
@@ -1114,7 +1007,7 @@ func (sess *Session) ResumePut(src, dst string) {
 		return
 	}
 	//生成临时文件
-	TempPath = path.Join(os.TempDir(), TempPath)
+	TempPath := path.Join(os.TempDir(), time.Now().String(), Temp)
 
 	file, err := os.OpenFile(TempPath, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0666)
 	defer file.Close()
@@ -1126,21 +1019,18 @@ func (sess *Session) ResumePut(src, dst string) {
 	_, err = write.WriteString(time.Now().Format("2006-01-02-15-04"))
 	if err != nil {
 		PrintErrorAndExit("write string error, error= %s", err)
-		return
 	}
 	_, err = write.Write(data)
 	if err != nil {
 		PrintErrorAndExit("write data to file error, error= %s", err)
-		return
 	}
 	_, err = write.WriteString("\n")
 	if err != nil {
 		PrintErrorAndExit("write string error, error= %s", err)
-		return
 	}
 	if err = write.Flush(); err != nil {
 		PrintErrorAndExit("write flush error, error= %s", err)
 	}
 
-	PrintErrorAndExit("ResumePut file %s to %s fail", src, dst)
+	PrintErrorAndExit("ResumePut file %s to %s fail", src, dst) //todo
 }
